fix: typing and chain not working
