+++ src/components/AssetManagementPanel/AssetManagementApplication/index.tsx
@@ -1,27 +1,19 @@
-import React, { useEffect, useState, useCallback } from "react";
+import React, { useState } from "react";
 import { useProviderContext } from "../../../common/contexts/provider";
 import { useTokenInformationContext } from "../../../common/contexts/TokenInformationContext";
 import { AssetManagementActions } from "../AssetManagementActions";
 import { AssetManagementForm } from "./../AssetManagementForm";
 import { AssetManagementTags } from "./../AssetManagementTags";
-import { getLogger } from "../../../utils/logger";
-import { useDispatch } from "react-redux";
-import { resetCertificateState } from "../../../reducers/certificate";
 interface AssetManagementApplicationProps {
-  tokenId: string;
   tokenRegistryAddress: string;
   setShowEndorsementChain: (payload: boolean) => void;
 }
 
-const { trace } = getLogger("component: assetmanagementapplication");
-
 export const AssetManagementApplication = ({
-  tokenId,
   tokenRegistryAddress,
   setShowEndorsementChain,
 }: AssetManagementApplicationProps) => {
   const {
-    initialize,
     approvedHolder,
     holder,
     approvedBeneficiary,
@@ -38,23 +30,9 @@ export const AssetManagementApplication = ({
     approveNewTransferTargetsState,
     transferToNewEscrow,
     transferToNewEscrowState,
-    resetStates: resetTokenInformationState,
   } = useTokenInformationContext();
   const [assetManagementAction, setAssetManagementAction] = useState(AssetManagementActions.None);
   const { upgradeProvider, account } = useProviderContext();
-  const dispatch = useDispatch();
-
-  const resetCertificateData = useCallback(() => dispatch(resetCertificateState()), [dispatch]);
-
-  useEffect(() => {
-    trace("initialise token information context");
-    initialize(tokenRegistryAddress, tokenId);
-    return () => {
-      trace("reseting token information on unmount");
-      resetTokenInformationState();
-      resetCertificateData();
-    };
-  }, [initialize, tokenId, tokenRegistryAddress, resetTokenInformationState, resetCertificateData]);
 
   const onSurrender = () => {
     transferTo(tokenRegistryAddress);
