+++ src/components/CertificateViewer.tsx
@@ -1,6 +1,6 @@
 import { VerificationFragment } from "@govtechsg/oa-verify";
 import { getData, WrappedDocument } from "@govtechsg/open-attestation";
-import React, { useCallback, useState } from "react";
+import React, { useCallback, useState, useEffect } from "react";
 import { Tab } from "react-bootstrap";
 import { getDocumentId, getTokenRegistryAddress } from "../common/utils/document";
 import { TemplateProps } from "./../types";
@@ -15,6 +15,12 @@ import { ErrorBoundary } from "./ErrorBoundary";
 import { ModalDialog } from "./ModalDialog";
 import { MultiButtons } from "./MultiButtons";
 import { TabPaneAttachments } from "./TabPaneAttachments";
+import { getLogger } from "../utils/logger";
+import { useDispatch } from "react-redux";
+import { resetCertificateState } from "../reducers/certificate";
+import { useTokenInformationContext } from "../common/contexts/TokenInformationContext";
+
+const { trace } = getLogger("component: certificateviewer");
 
 interface CertificateViewerProps {
   document: WrappedDocument;
@@ -42,6 +48,24 @@ export const CertificateViewer = ({
   const originalData = getData(document);
   const attachments = originalData?.attachments;
   const hasAttachments = attachments && attachments.length > 0;
+  const { initialize, resetStates: resetTokenInformationState } = useTokenInformationContext();
+  const dispatch = useDispatch();
+
+  const resetCertificateData = useCallback(() => dispatch(resetCertificateState()), [dispatch]);
+
+  /* 
+  initialise the meta token information context when new tokenId 
+  and tokenRegistryAddress provided and clean up the context when certificate viewer unmounts
+  */
+  useEffect(() => {
+    trace("initialise token information context");
+    initialize(tokenRegistryAddress, tokenId);
+    return () => {
+      trace("reseting token information on unmount");
+      resetTokenInformationState();
+      resetCertificateData();
+    };
+  }, [tokenId, tokenRegistryAddress, resetCertificateData, resetTokenInformationState, initialize]);
 
   const childRef = React.useRef<{ print: () => void }>();
 
@@ -79,7 +103,6 @@ export const CertificateViewer = ({
         <DocumentStatus verificationStatus={verificationStatus} />
         {tokenRegistryAddress && (
           <AssetManagementApplication
-            tokenId={tokenId}
             tokenRegistryAddress={tokenRegistryAddress}
             setShowEndorsementChain={setShowEndorsementChain}
           />
